created slinter rules, refactored the code to meet the rules and created custom markers
