+++ frontend/src/components/api.js
@@ -1,7 +1,7 @@
 export function doFetch(url, method) {
-    return fetch(url, { method: method, headers: { 'Content-Type': 'application/json' } })
+  return fetch(url, { method, headers: { 'Content-Type': 'application/json' } });
 }
 
 export function doFetchBody(url, method, body) {
-    return fetch(url, { method: method, headers: { 'Content-Type': 'application/json' }, body: JSON.stringify(body) })
+  return fetch(url, { method, headers: { 'Content-Type': 'application/json' }, body: JSON.stringify(body) });
 }
